requirements.txt
fastapi==0.115.10
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic
asyncpg
psycopg[binary]
pydantic  # Pydantic v2.x
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
python-decouple==3.8
redis==5.0.1
celery==5.3.4
stripe==7.8.0
elasticsearch==8.11.0
boto3==1.34.0
emails==0.6
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
factory-boy==3.3.0
faker==20.1.0
slowapi==0.1.9
prometheus-client==0.19.0
